feat: implement alternative Table of Contents widget and update references
