+++ linux/flutter/ephemeral/.plugin_symlinks/url_launcher_linux
@@ -1,0 +1,1 @@
+/Users/dartaan/.pub-cache/hosted/pub.dev/url_launcher_linux-3.2.1/
